basic Validation mode
It is a very basic version of Validation mode. It inputs the calibrated parameters from Bayesian (in json). Then it selects the set with the highest likelihood and runs the model based on this selected set. Finally, it saves the simulated values generated by the model in Validation.csv.

Next: we should compare the simulated values (from Validation run) with the measured values