added Latin Hypercube
It now samples 10 sets (the number is up to us, we should discuss it).

However, it samples the normal distribution of likelihoods. But I am not sure whether it's correct, as the best likelihoods are the highest ones?

Also, for now, the respiration plots are generated for every set of parameters (will be changed to means over every treatment) and you can see a huuuuge differences between sets