Normal and Bayesian mode update for Jilkova2022
both Normal and Bayesian modes now work for both datasets - Jilkova2022 and Jilkova2024, therefore I got rid of mode "Jilkova2022":
Edited how respiration plot is made, so that it is versatile.

For Jilkova2022 dataset, I use soil-dervived respiration and substrate-derived respiration for calibration (the ones in plot), instead of total respiration and resp_sub, so that we have the same units for those. Todo: change this also for Bayesian_input2024, and in both inputs do the same switch also for pools (DOM, POM, MAOM, fungi, bacteria)