big edit: normal run with validation as function, including 1:1 plots
For now I cancelled "Validation mode", we have Normal mode, in which validation is also carried out (comparison with measured values). I wrote the normal mode as a function so that it can be also called at the end of Bayesian to automatically try out the best parameter set. There are new 1:1 plots showing modelled against measured for all variables that we also calibrate towards. For each variable model performance metrics (RMSE, bias, EF) are calculated and a mean EF is calculated as overall model performance score.